requirements.txt
fastapi
mysql-connector-python
pydantic
python-dotenv
SQLAlchemy
uvicorn
jinja2
PyJWT
pymysql
python-multipart
